2023/2/1

+++ data.sh
@@ -1,5 +1,6 @@
 #!/bin/sh 
-DATASET= ""
+# eurlex-4k, wiki10-31k, amazon-3m,amazon-670k,wiki-500k
+DATASET= 
 if [ $# == 0 ];then
     DATASET="eurlex-4k"
 else
@@ -7,11 +8,17 @@ else
 fi
 echo ${DATASET}
 
-# 下载数据集downloading dataset
-wget -nv -nc https://archive.org/download/pecos-dataset/xmc-base/${DATASET}.tar.gz -P ./dataset/
-#解压 uncompress tar file
-tar -zvxf ./dataset/${DATASET}.tar.gz -C ./dataset/
-# file归档
-mv  ./dataset/xmc-base/${DATASET} ./dataset
-rm -r ./dataset/xmc-base
-rm ./dataset/${DATASET}.tar.gz
+if [ ! -d "./dataset/${DATASET}" ];then
+    # 下载数据集downloading dataset
+    wget -nc https://archive.org/download/pecos-dataset/xmc-base/${DATASET}.tar.gz -P ./dataset/
+    #解压 uncompress tar file
+    tar -zvxf ./dataset/${DATASET}.tar.gz -C ./dataset/
+    # file归档
+    mv  ./dataset/xmc-base/${DATASET} ./dataset
+    rm -r ./dataset/xmc-base
+    rm ./dataset/${DATASET}.tar.gz 
+ else
+ echo "${DATASET}文件存在"
+fi
+
+
